Added a new scihub url to the list
Also added a note for the researcher to go to researchgate or the publisher's website if the PDF is not found. Not everything is on sci-hub

Note: I'm trying to keep the number of urls low to not make the user wait too long if the PDF is not found on the first one.